00_assignments_class/09_assignment/receipt-interpreter/requirements.txt
streamlit
streamlit
python-dotenv
google-generativeai
openai-agents
Pillow